reorganise modules and tools
