+++ requirements.txt
@@ -4,4 +4,4 @@ pandas==2.2.2
 python-dateutil==2.8.2
 requests==2.25.0
 tabulate==0.9.0
-pytables==3.10.1
+tables==3.10.1
