+++ setup.sh
@@ -1,5 +1,6 @@
 PERFORMANCE_TEST_PATH=`realpath $BASH_SOURCE | xargs dirname`
 export PERFORMANCE_TEST_PATH
+export PYTHONPATH=$PERFORMANCE_TEST_PATH/src:$PYTHONPATH
 export PATH=$PERFORMANCE_TEST_PATH/tools:$PATH
 export PATH=$PERFORMANCE_TEST_PATH/scripts:$PATH
 export PATH=$PERFORMANCE_TEST_PATH/tests:$PATH
